requirements.txt
pyspark==3.5.1
apache-sedona[spark]==1.6.0
requests~=2.32.3
